+++ src/internal/getCaseResult.ts
@@ -1,4 +1,5 @@
 import { Event } from 'benchmark'
+import * as median from 'stats-median'
 import { CaseResult } from './common-types'
 
 type GetCaseResult = (event: Event) => CaseResult
@@ -19,6 +20,19 @@ const getCaseResult: GetCaseResult = (event) => {
     },
     samples: target.stats.sample.length,
     promise: target.defer,
+    details: {
+      min: Math.min(...target.stats.sample),
+      max: Math.max(...target.stats.sample),
+      mean: target.stats.mean,
+      median: median.calc(target.stats.sample),
+      standardDeviation: target.stats.deviation,
+      marginOfError: target.stats.moe,
+      relativeMarginOfError: target.stats.rme,
+      standardErrorOfMean: target.stats.sem,
+      sampleVariance: target.stats.variance,
+      sampleResults: target.stats.sample,
+    },
+    completed: target.stats.sample.length > 0,
   }
 }